+++ package.json
@@ -7,7 +7,9 @@
     "test:vitest": "vitest --passWithNoTests --reporter verbose",
     "generate:pdfs": "tsx ./src/bin/generate-pdfs.ts",
     "dev": "nodemon --watch src --exec pnpm run generate:pdfs",
-    "build": "tsx ."
+    "build": "tsx .",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "keywords": [],
   "author": "",
@@ -21,12 +23,22 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.7.1",
+    "@chromatic-com/storybook": "^1.3.3",
+    "@storybook/addon-essentials": "^8.0.9",
+    "@storybook/addon-interactions": "^8.0.9",
+    "@storybook/addon-links": "^8.0.9",
+    "@storybook/addon-onboarding": "^8.0.9",
+    "@storybook/blocks": "^8.0.9",
+    "@storybook/react": "^8.0.9",
+    "@storybook/react-vite": "^8.0.9",
+    "@storybook/test": "^8.0.9",
     "@types/node": "^20.12.7",
     "@types/react": "^18.2.79",
     "@types/react-dom": "^18.3.0",
     "@types/yargs": "^17.0.32",
     "glob": "^10.3.12",
     "nodemon": "^3.1.0",
+    "storybook": "^8.0.9",
     "tsx": "^4.7.3",
     "typescript": "^5.4.5",
     "vitest": "^1.5.2",
