+++ src/templates/DefaultArgs.tsx
@@ -1,0 +1,59 @@
+import type { BoletoProps } from "./InputSchema";
+
+export const defaultArgs: BoletoProps = {
+  barcodeData: "23790.34404 20000.301430 99001.327503 9 97120000009332",
+  scannerBarcode: "23790.34404 20000.301430 99001.327503 9 97120000009332",
+  paymentPlace:
+    "Pagável preferencialmente na rede Bradesco ou Bradesco Expresso.",
+  beneficiary: "UNICRED FLORIANÓPOLIS - CNPJ: 074.064.502/0001-12",
+  beneficiaryAddress:
+    "Rua Tenete Silveira, 315 - Centro - Florianópolis - SC  - CEP 88010-301",
+  instructions:
+    "Após o vencimento cobrar multa de 2,00% , mais juros ao mes de 1,00%.",
+  agency: "7506",
+  agencyDigit: "0",
+  account: "54291",
+  accountDigit: "1",
+  expirationDay: new Date(2017, 7, 30),
+  documentDate: new Date(2017, 7, 18),
+  processingDate: new Date(2017, 7, 18),
+  card: "09",
+  documentNumber: "42493",
+  formattedOurNumber: "09/19000001208-0",
+  formattedValue: "R$ 9,90",
+  documentType: "DS",
+  accept: "N",
+  currencyType: "Real (R$)",
+  amount: " ",
+  valueOf: " ",
+  discountValue: " ",
+  otherDiscounts: " ",
+  feeValue: " ",
+  otherFees: " ",
+  chargeValue: " ",
+  qrCodeData:
+    "00020126580014br.gov.bcb.pix01369e94d622-aecd-4f41-9b0e-5e9342465383520400005303986540510.005802BR5910Flop_Robux6009Sao_Paulo62290525617a30238ffb442e99a6a0c266304F356",
+  payer: {
+    name: "Anita Albuquerque",
+    registerNumber: "221.412.772-05",
+    street: "Rua Maria Gertrudes Coelho",
+    number: "827",
+    complement: " ",
+    district: "Estrada Nova",
+    city: "Divinópolis",
+    state: "MG",
+    postalCode: "35500-700",
+  },
+  guarantor: {
+    name: "ACME Telecomunicações Ltda",
+    registerNumber: "074.064.502/0001-12",
+    street: "Servidão",
+    number: "439",
+    district: "Estrada Nova",
+    complement: " ",
+    city: "Jaraguá do Sul",
+    state: "SC",
+    postalCode: "89254-375",
+  },
+  isServer: true,
+};
